rag-backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.45.0
anyio==4.12.1
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cohere==5.20.5
colorama==0.4.6
cryptography==46.0.5
deprecation==2.1.0
distro==1.9.0
eval_type_backport==0.3.1
fastavro==1.12.1
filelock==3.21.0
fsspec==2026.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.190.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface_hub==1.4.1
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.1
invoke==2.2.1
jiter==0.13.0
Markdown==3.10.2
markdown-it-py==4.0.0
mdurl==0.1.2
mistralai==1.12.1
mmh3==5.2.0
multidict==6.7.1
numpy==2.4.2
openai==2.20.0
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
packaging==26.0
pillow==12.1.1
postgrest==2.28.0
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyiceberg==0.11.0
PyJWT==2.11.0
PyMuPDF==1.27.1
pyparsing==3.3.2
PyQt6==6.10.2
PyQt6-Qt6==6.10.2
PyQt6_sip==13.11.0
pyroaring==1.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
realtime==2.28.0
regex==2026.1.15
requests==2.32.5
rich==14.3.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
storage3==2.28.0
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.28.0
supabase-auth==2.28.0
supabase-functions==2.28.0
tenacity==9.1.4
tiktoken==0.12.0
tokenizers==0.22.2
tqdm==4.67.3
typer==0.23.0
typer-slim==0.23.0
types-requests==2.32.4.20260107
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
websockets==15.0.1
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
google-genai>=1.0,<2
fastapi>=0.115,<1
uvicorn[standard]>=0.34,<1
python-multipart>=0.0.18
slowapi>=0.1.9
limits>=3.13.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
